. r some minor upds
