--- packages/ispect/lib/src/features/json_viewer/pretty_json_view.dart
@@ -1,2 +1,0 @@
-export 'src/models/json_node.dart';
-export 'src/theme/json_view_theme.dart';
