+++ packages/ispect/lib/src/features/json_viewer/store.dart
